refactor: request, response 변수명 변경
* Request -> Req
* Response -> Res
