+++ denstiny/sql-db/src/main/java/com/example/board_image/BoardImageEntity.java
