+++ denstiny/sql-db/src/main/java/com/example/board_image/BoardImageRepository.java
